requirements.txt
asgiref==3.11.0
blockfrost-python==0.6.0
boto3==1.40.59
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
emoji==2.15.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
mysqlclient==2.2.6
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
whitenoise==6.11.0
PyMySQL==1.1.1